Writing new script for executing JS animations
Use of jQuery/JS enables incorporation of advanced features, such as more simplified fallback animations when a page is viewed multiple times over the same session. Also avoids limitations associated with CSS3 animations and the status of the CSS editor in the WordPress configuration, which precludes use of the animation-fill-mode property.