Fix bugs in TextAutoResizingElem class
Remove unnecessary call to jQuery in function initArticleHeaderText(…); also refactor method TextAutoResizers.initTextAutoResizing to resolve a strange fatal JS error that is triggered by utilizing a regular expression literal on pages that have a gravity form with a rich text editor. I suspect it is related to the fact that RegEx literals are compiled only on script loading, and rich text editors use an iframe to embed WordPress' TinyMCE rich editor into the form.