Fix highlighting of rich text fields in gravity forms
Update the function hghlghtRqrdRchTxtEdtrs in jQuery.forms.js, introduce an delay before executing the code that finds the iframes in which rich text editors are contained and updates the styling of the fields. This is delay is necessary because Gravity Forms plugin utilizes JavaScript to dynamically replace a textarea element with an iframe in which the rich text editing interface is loaded. Fixes #5.