Feat: Dev tools script → WSUWP page data
Update «printWpPageDetails.js» to automatically navigate to the first page of the "All Pages" WSUWP interface, if needed, when the script is run for the first time during a browsing session. Also, improve how dates and times are scanned by the script so they are easier to work with in spreadsheet software.