Feat: Dev tools script → Scan latest WSU majors
Write «scanLatestListOfWsuMajors.devTools.js» as a utility script for deployment in a web browser's developer tools that will automatically scan the WSU website for the latest official lists of the university's majors and minors. Write the script to take advantage of the session storage API to keep track of scan results and location information as the script moves the browser from page to page.