Feat: WDS cons6t extractor → Improve clipboard output
Improve the mechanism in «DevTools/extractWdsConstituentsFromGH.js» to copy output to the clipboard. Use a Promise to wait for the user to switch focus from the DevTools console back to the document so the Clipboard API can be successfully used to output extracted data. Alert the user once the clipboard operation has completed.